src/test/javascript/spec/app/entities/movimiento/movimiento.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import * as moment from 'moment';
import { DATE_TIME_FORMAT } from 'app/shared/constants/input.constants';
import { MovimientoService } from 'app/entities/movimiento/movimiento.service';
import { IMovimiento, Movimiento } from 'app/shared/model/movimiento.model';

describe('Service Tests', () => {
  describe('Movimiento Service', () => {
    let injector: TestBed;
    let service: MovimientoService;
    let httpMock: HttpTestingController;
    let elemDefault: IMovimiento;
    let expectedResult: IMovimiento | IMovimiento[] | boolean | null;
    let currentDate: moment.Moment;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(MovimientoService);
      httpMock = injector.get(HttpTestingController);
      currentDate = moment();

      elemDefault = new Movimiento(0, currentDate, 'AAAAAAA', 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign(
          {
            whenDate: currentDate.format(DATE_TIME_FORMAT),
          },
          elemDefault
        );

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a Movimiento', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
            whenDate: currentDate.format(DATE_TIME_FORMAT),
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            whenDate: currentDate,
          },
          returnedFromService
        );

        service.create(new Movimiento()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a Movimiento', () => {
        const returnedFromService = Object.assign(
          {
            whenDate: currentDate.format(DATE_TIME_FORMAT),
            description: 'BBBBBB',
            cantidad: 1,
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            whenDate: currentDate,
          },
          returnedFromService
        );

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of Movimiento', () => {
        const returnedFromService = Object.assign(
          {
            whenDate: currentDate.format(DATE_TIME_FORMAT),
            description: 'BBBBBB',
            cantidad: 1,
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            whenDate: currentDate,
          },
          returnedFromService
        );

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Movimiento', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
